Add exception handling for entity not found and unique foreign key violations

+++ src/main/java/dev/wannabe/reciperepository/service/FinalProductService.java
@@ -1,6 +1,7 @@
 package dev.wannabe.reciperepository.service;
 
 
+import jakarta.persistence.EntityNotFoundException;
 import dev.wannabe.reciperepository.model.FinalProduct;
 import dev.wannabe.reciperepository.repository.FinalProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,8 +32,19 @@ public class FinalProductService {
     public FinalProduct save(FinalProduct finalProduct) {
         return finalProductRepository.save(finalProduct);
     }
+    public FinalProduct update(FinalProduct finalProduct) {
+        if(!finalProductRepository.existsById(finalProduct.getId())){
+            throw new EntityNotFoundException("FinalProduct on id " + finalProduct.getId() + " not found");
+        }
+        return finalProductRepository.save(finalProduct);
+    }
 
     public void deleteById(Long id) {
+
+        if(!finalProductRepository.existsById(id)){
+            throw new EntityNotFoundException("FinalProduct on id " + id + " not found");
+        }
+
         finalProductRepository.deleteById(id);
     }
