+++ src/main/java/dev/wannabe/reciperepository/service/RecipeService.java
@@ -3,6 +3,7 @@ package dev.wannabe.reciperepository.service;
 
 import dev.wannabe.reciperepository.model.Recipe;
 import dev.wannabe.reciperepository.repository.RecipeRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -29,8 +30,17 @@ public class RecipeService {
     public Recipe save(Recipe recipe) {
         return recipeRepository.save(recipe);
     }
+    public Recipe update(Recipe recipe) {
+        if(!recipeRepository.existsById(recipe.getId())){
+            throw new EntityNotFoundException("Recipe on id " + recipe.getId() + " not found");
+        }
+        return recipeRepository.save(recipe);
+    }
 
     public void deleteById(Long id) {
+        if(!recipeRepository.existsById(id)){
+            throw new EntityNotFoundException("Recipe on id " + id + " not found");
+        }
         recipeRepository.deleteById(id);
     }
