+++ src/main/java/dev/wannabe/reciperepository/service/SupplierService.java
@@ -3,6 +3,7 @@ package dev.wannabe.reciperepository.service;
 
 import dev.wannabe.reciperepository.model.Supplier;
 import dev.wannabe.reciperepository.repository.SupplierRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -32,7 +33,17 @@ public class SupplierService {
         return supplierRepository.save(supplier);
     }
 
+    public Supplier update(Supplier supplier) {
+        if(!supplierRepository.existsById(supplier.getId())){
+            throw new EntityNotFoundException("Supplier on id " + supplier.getId() + " not found");
+        }
+        return supplierRepository.save(supplier);
+    }
+
     public void deleteById(Long id) {
+        if(!supplierRepository.existsById(id)){
+            throw new EntityNotFoundException("Supplier on id " + id + " not found");
+        }
         supplierRepository.deleteById(id);
     }
 }
