+++ src/main/java/dev/wannabe/reciperepository/service/ToolService.java
@@ -3,6 +3,7 @@ package dev.wannabe.reciperepository.service;
 
 import dev.wannabe.reciperepository.model.Tool;
 import dev.wannabe.reciperepository.repository.ToolRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -22,6 +23,13 @@ public class ToolService {
         return toolRepository.save(tool);
     }
 
+    public Tool update(Tool tool) {
+        if(!toolRepository.existsById(tool.getId())){
+            throw new EntityNotFoundException("Tool on id " + tool.getId() + " not found");
+        }
+        return toolRepository.save(tool);
+    }
+
     public Tool findById(Long id) {
         return toolRepository.findById(id).orElse(null);
     }
@@ -31,6 +39,9 @@ public class ToolService {
     }
 
     public void deleteById(Long id) {
+        if(!toolRepository.existsById(id)){
+            throw new EntityNotFoundException("Tool on id " + id + " not found");
+        }
         toolRepository.deleteById(id);
     }
 }
